1.1.0+2025/07/04
Se finaliza el desarrollo de pruebas unitarias para service y controladores, se realiza documentación con swagger, y el DataLoader

+++ src/main/java/com/SAFE_Rescue/API_Recursos/repository/BomberoRepository.java
@@ -9,7 +9,7 @@ import org.springframework.stereotype.Repository;
  * Maneja operaciones CRUD desde la base de datos usando Jakarta
  */
 @Repository
-public interface BomberoRepository extends JpaRepository<Bombero, Long>{
+public interface BomberoRepository extends JpaRepository<Bombero, Integer>{
 
 }
